fix(dsv41): preserve native FlashInfer routed weight ABI
(cherry picked from commit 0ccc40dfd8ceb0fbe56cb68e1dce2890c21cc8bd)
(cherry picked from commit 6a899fd8296f3b8f6e1a41ace558ed0f7d8b5057)

+++ python/sglang/srt/layers/quantization/mxfp4_flashinfer_trtllm_moe.py
@@ -471,7 +471,11 @@ class Mxfp4FlashinferTrtllmMoEMethod:
             enable_pdl=trtllm_moe_enable_pdl(num_tokens),
         )
         if defer_finalize:
-            output = _make_deferred_finalize_output(result, top_k=topk_ids.shape[1])
+            output = _make_deferred_finalize_output(
+                result,
+                top_k=topk_ids.shape[1],
+                expected_expert_weights_dtype=topk_weights.dtype,
+            )
         else:
             output = result[0]
